+++ src/Model/Table/AppTable.php
@@ -2,6 +2,7 @@
 
 namespace MeTools\Model\Table;
 
+use Cake\ORM\Query;
 use Cake\ORM\Table;
 use MeTools\Model\Validation\AppValidator;
 
@@ -20,5 +21,16 @@ abstract class AppTable extends Table
         parent::initialize($config);
 
         $this->_validatorClass = AppValidator::class;
+    }
+
+    /**
+     * `active` find method
+     * @param \Cake\ORM\Query $Query The query builder
+     * @return \Cake\ORM\Query The query builder
+     * @see \Cake\ORM\Table::find() for options to use for the find
+     */
+    public function findActive(Query $Query): Query
+    {
+        return $Query->where(['active' => true]);
     }
 }