+++ tests/TestCase/Model/Table/AppTableTest.php
@@ -25,6 +25,11 @@ use MeTools\TestSuite\TestCase;
 class AppTableTest extends TestCase
 {
     /**
+     * @var array<string>
+     */
+    protected $fixtures = ['app.Users'];
+
+    /**
      * @var \MeTools\Model\Table\AppTable|\PHPUnit\Framework\MockObject\MockObject
      */
     protected AppTable $AppTable;
@@ -48,5 +53,19 @@ class AppTableTest extends TestCase
     {
         $this->AppTable->initialize([]);
         $this->assertInstanceOf(AppValidator::class, $this->AppTable->getValidator());
+    }
+
+    /**
+     * @test
+     * @uses \MeTools\Model\Table\AppTable::findActive()
+     */
+    public function testFindActive(): void
+    {
+        //Uses a table that extends `AppTable` and has the `active` field
+        $Query = $this->getTableLocator()->get('Users')->find('active');
+
+        $this->assertGreaterThan(0, $Query->count());
+        $this->assertStringEndsWith('FROM users Users WHERE active = :c0', $Query->sql());
+        $this->assertTrue($Query->getValueBinder()->bindings()[':c0']['value']);
     }
 }