+++ CHANGELOG.md
@@ -29,7 +29,8 @@ All the code has been optimized overall and has been concentrated in fewer metho
   `text-decoration-` class is already present;
 * added the `AbstractPerson` entity class;
 * added basic abstract `AppValidator` class with some general rules;
-* added basic `AppController` and `AppTable` abstract classes and `View` classe;
+* added basic abstract `AppTable` class, with a default validator and a `findActive()` method;
+* added basic `AppController` abstract class and `View` classe;
 * added the `AddButtonClassesTrait` with the `addButtonClasses()` method;
 * the `OptionsParser` is now deprecated and will be removed in a later release. The same goes for the global
   `optionsParser()` function;